Fixed LevelBorwser Pagination for duplicate db field values

+++ Assets/Scripts/Data/FirebaseLevelData.cs
@@ -10,7 +10,7 @@ public class FirebaseLevelData : MonoBehaviour {
     public string AuthorID;
     public string LevelID;
     public int QtyRatings;
-    public int UserRating;
+    public float UserRating;
     public int Difficulty;
     public string LevelDataURL;
